backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pydantic[email]
pydantic-settings
python-dotenv
google-generativeai